fix(ai-gateway): avoid model registry dependency cycle
